AMB-1260: Fixed Spring Security failing to authenticate due to missing UserInfo.
This fix relies on providing `application.yml` with `user-info-uri` and `userNameAttribute` values, we're passing the `/userinfo` endpoint for this purpose - but this only works for CIS2, since NHS Login doesn't have this endpoint available.

For NHS Login we'll likely have to grab the ID token from Apigee somehow and create a custom AuthenticationProvider similar to `OidcAuthorizationCodeAuthenticationProvider` to grab the user properties Spring Security needs from this ID token.